requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.0.1
pillow==10.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
PyJWT==2.8.0
python-dateutil==2.9.0.post0
reportlab==4.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
